calculadora

+++ calculadora.py
@@ -1,8 +1,5 @@
-numero1 = int(input("numero 1:"))
-numero2 = int(input("numero 2:"))
+numero1 = float(input("numero 1:"))
+numero2 = float(input("numero 2:"))
 print("A subtração de", numero1,"e", numero2, "é", numero1-numero2)
 print("A multiplicação do", numero1,"e do",numero2,"é",numero2*numero1)
-#numero1 = float(input("numero 1:"))
-#numero2 = float(input("numero 2:"))
-#print("A soma de", numero1,"e", numero2, "é", numero1+numero2)
 #float = numeros quebrados, int = numeros inteiros
